2-Backend/eventrackers/requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
asgiref==3.3.1; python_version >= '3.5'
colorama==0.4.4; platform_system == 'Windows'
django==3.1.5
djangorestframework==3.12.2
pillow==8.1.0
pytz==2020.5
qrcode==6.0
six==1.15.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.4.1; python_version >= '3.5'
